ported RCS from python to java

+++ gopiandcode-xcs/src/main/java/com/gopiandcode/lcs/BinaryClassifierTestRunner.java
@@ -79,7 +79,7 @@ public class BinaryClassifierTestRunner {
 
             i++;
        }
-        return correctlyPredicted/count;
+        return (double)correctlyPredicted/(double)count;
     }
 
     public void setShouldReset(boolean shouldReset) {
